requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
jinja2==3.1.2
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.7
jinja2-time==0.2.0
python-multipart==0.0.6
